Support step-function fee schedules
Note that flat fees make sense only for linear fee schedules because
step-function schedules are a form of flat fees.
Also, the previous commit left Carpark.Fee() in place, not because it
will remain part of the Carpark class, but because it is not yet clear
where this operation should reside.  Probably, it will end up in the
Charge class.