Use enums on interfaces
Some interface messages could carry either an enumerated data type or
the string equivalent.  For example,
ControlRoomOperator:ActivateEntryStand carries two such parameters,
Barrier and Ticket.  This message is sent from the CarparkControl
component to the Operator UI, so it is tempting to send a string.
However, the Operator UI is wrapped by a shell component which receives
all incoming messages bound for the Operator UI.
For each incoming message, this shell component provides:
A function, the shell, including parameter declarations, of which is
generated and the body of which is hand coded to make the connection to
the externally produced UI.
A port activity, expressed in OAL, which invokes the function mentioned
above, passing any message parameters.
Accordingly, the interface messages carry enumerated data types.  Any
conversion from enums to strings, or perhaps graphical representations,
is then left to the shell component or the external code.